make sec uint32_t

+++ fw-mm/timer2.h
@@ -5,6 +5,6 @@
  *
  */
 
-extern volatile uint8_t sec;
+extern volatile uint32_t sec;
 
 void timer2_init(void);
